moved image buffer to seprate package because cleanliness
